[PATCH] /dev/random: Remove RNDGETPOOL ioctl
Recently, someone has kvetched that RNDGETPOOL is a "security
vulnerability".  Never mind that it is superuser only, and with superuser
privs you could load a nasty kernel module, or read the entropy pool out of
/dev/mem directly, but they are nevertheless still spreading FUD.

In any case, no one is using it (it was there for debugging purposes only),
so we can remove it as dead code.

Signed-off-by: Andrew Morton <akpm@osdl.org>
Signed-off-by: Linus Torvalds <torvalds@osdl.org>

BKrev: 412b8ca2lcWCKDQrRiKH3VZHvek3Tg