Merge bk://kernel.bkbits.net/davem/net-2.6 into ppc970.osdl.org:/home/torvalds/v2.6/linux
2004/08/24 11:35:14-07:00 sds
[SELINUX]: Fix bugs introduced by skb_header_pointer() changes.

Lines assigning initial value to 'ret' were removed
erroneously.

Signed-off-by: Stephen Smalley <sds@epoch.ncsc.mil>
Signed-off-by: David S. Miller <davem@redhat.com>

2004/08/24 11:32:44-07:00 yoshfuji
[IPSEC]: Add SCTP to xfrm_flowi_{sport,dport}()

Signed-off-by: HIDEAKI Yoshifuji <yoshfuji@linux-ipv6.org>
Signed-off-by: David S. Miller <davem@redhat.com>

2004/08/24 11:30:31-07:00 herbert
[IPSEC]: Set TTL from route.

Here is the promised patch that sets the TTL from the route parameter.
I decided against adding an option to inherit the TTL like IPIP/GRE
as I think that it doesn't really make sense with IPsec.  But it
can be easily added later if someone needs it.

This isn't completely right when nested tunnels are involved.  The
TTL for intervening tunnels should be set from the routes to the
intervening nodes.  But fixing that involves using information that
isn't currently in the bundle.  I'll revisit this once the MTU stuff
is fixed since that'll also involving adding the intervening routes
to the bundle.

Signed-off-by: Herbert Xu <herbert@gondor.apana.org.au>
Signed-off-by: David S. Miller <davem@redhat.com>

2004/08/24 11:28:16-07:00 ajgrothe
[CRYPTO]: Add Whirlpool digest algorithm.

Given the recent potential weaknesses in the SHA and MD families,
I thought it might not be a bad idea to include another hash/digest
algorithm in the kernel.

So here is Whirlpool.  I chose it for a couple of reasons.

o - It is by the same people who did Khazad. I feel pretty good about their work.
o - It has been evaluated by NESSIE
  https://www.cosic.esat.kuleuven.ac.be/nessie/reports/phase1/sagwp3-037_1.pdf
o - NESSIE has accepted it as one of the cryptographic primitives
o - It will be part of an ISO standard in the revised ISO/IEC 10118-3:2003(E) standard, thanks to
NESSIE
o - It is patent free and has an implementation in the public domain.

Signed-off-by: Aaron Grothe <ajgrothe@yahoo.com>
Signed-off-by: James Morris <jmorris@redhat.com>
Signed-off-by: David S. Miller <davem@redhat.com>

2004/08/24 11:23:53-07:00 lkml
[NETFILTER]: Missing netfilter_ipv4.c include in conntrack proto code.

Signed-off-by: Felipe Alfaro Solana <lkml@felipe-alfaro.com>
Signed-off-by: David S. Miller <davem@redhat.com>

2004/08/24 11:21:01-07:00 yoshfuji
[IPV6]: Fix device handling in ip6_route_add().

Signed-off-by: HIDEAKI Yoshifuji <yoshfuji@linux-ipv6.org>
Signed-off-by: David S. Miller <davem@redhat.com>

BKrev: 412b9a7565GgC_DBGvxW4-HXSDa1cw