Merge davem@nuts.davemloft.net:/disk1/BK/net-2.6 into kernel.bkbits.net:/home/davem/net-2.6
2004/08/25 20:20:04-07:00 laforge
[NETFILTER]: Fix ip_nat_find_helper() locking.

Signed-off-by: Harald Welte <laforge@netfilter.org>
Signed-off-by: David S. Miller <davem@redhat.com>

2004/08/25 16:35:34-07:00 shemminger
[BRIDGE]: Fix oops when mangling and brouting and tcpdumping packets

The ebtables brouting chain, traversed through the call
br_should_route_hook(), can alter a packet. The redirect target
does this, f.e., to change the MAC destination.

Bart discovered this and proposed a patch; this is a revised version.
This version cleans up the handle_bridge code in net/core/dev.c as well
as getting rid of extra rcu_read_lock and only does the br_port checking
once.

Signed-off-by: Stephen Hemminger <shemminger@osdl.org>
Signed-off-by: David S. Miller <davem@redhat.com>

2004/08/25 15:58:31-07:00 davem
Merge nuts.davemloft.net:/disk1/BK/network-2.6
into nuts.davemloft.net:/disk1/BK/net-2.6

2004/08/24 17:20:56-07:00 tgraf
[NET]: Device mtu/txqlen/ifmap via rtnetlink.

Signed-off-by: Thomas Graf <tgraf@suug.ch>
Signed-off-by: David S. Miller <davem@redhat.com>

2004/08/24 17:10:20-07:00 shemminger
[NET]: Another cleanup in netif_receive_skb()

Move rcu_read_lock up a little, since it needs to be
done in both branches anyway. Also whitespace fix.

Signed-off-by: Stephen Hemminger <shemminger@osdl.org>
Signed-off-by: David S. Miller <davem@redhat.com>

2004/08/24 17:08:33-07:00 shemminger
[NET]: deliver_skb() cleanup

Cleanup of deliver_skb: get rid of unused argument and use it
in the NET_CLS_ACT hook.

Signed-off-by: Stephen Hemminger <shemminger@osdl.org>
Signed-off-by: David S. Miller <davem@redhat.com>

2004/08/24 17:03:36-07:00 herbert
[NET]: Use pskb_expand_head() instead of skb_copy() in skb_checksum_help().

Here is the patch that you wanted to shoot holes at :)

The idea is simple.  None of the callers of skb_checksum are passing it
skb's which are shared.  They may be cloned however.  But the application
checksum is always in the skb header so there is no need to linearise it.

Supposing all these assumptions are correct, then we can avoid the overhead
of skb_copy() and get away with pskb_expand_head().

If the assumption is wrong, we will find out because
pskb_expand_head() will BUG() if skb_shared() is true.

Signed-off-by: Herbert Xu <herbert@gondor.apana.org.au>
Signed-off-by: David S. Miller <davem@redhat.com>

BKrev: 412dbc81VVv4iTAB835wWSOBfvccMw