[PATCH] parport: NetMOS 9805 interface
Add support for netmos devices to the parallel port driver.

NetMOS 9805 support is already in the kernel, this patch adds the support for
the missing 9735,9855,9755 and 9715 chips.

And another remark: The 9735 and 9835 seem to be chips with serial *and*
parallel interfaces, so I suppose they are already claimed somewhere in the
serial driver.  I don't know whether this causes any problems.  I'm sorry that
I can't test, I've only a 9805 here.  Any idea how these "dual" chips have to
be handled by the kernel?

Signed-off-by: Andrew Morton <akpm@osdl.org>
Signed-off-by: Linus Torvalds <torvalds@osdl.org>

BKrev: 412f716aZjp3d0BilzBuNsejXLCdUw