Merge bk://kernel.bkbits.net/davem/sparc-2.6 into ppc970.osdl.org:/home/torvalds/v2.6/linux
2004/08/27 18:55:48-07:00 davem
[CREDITS]: Update my entry.

2004/08/27 18:48:38-07:00 davem
[SPARC64]: Update defconfig.

2004/08/27 18:33:58-07:00 davem
[AMD7930]: Fix kcalloc() args.

2004/08/27 17:08:47-07:00 davem
Merge nuts.davemloft.net:/disk1/BK/sparcwork-2.6
into nuts.davemloft.net:/disk1/BK/sparc-2.6

2004/08/27 16:19:23-07:00 davem
[MAINTAINERS]: Update my email contact info.

2004/08/27 16:12:23-07:00 davem
[SPARC64]: Fix delay with HZ==1000.

When I moved sparc64 over to HZ==1000 this added some
problems to the udelay() handling.  Specifically, with
slower cpus we could now get underflows to zero for
things like udelay(1) due to the order of multiplies
and shifts.

Fix this, and move it out to arch/sparc64/lib/delay.c
so it is easier to tinker with this in the future and
also to optimize away one of the multiplies for the
constant delay case just like other platforms do.

Signed-off-by: David S. Miller <davem@davemloft.net>

2004/08/27 13:45:44-07:00 davem
[UACCESS]: Fix typo in generic __get_user_unaligned().

Signed-off-by: David S. Miller <davem@redhat.com>

BKrev: 412fecb3B1sAtfeQlV492xbT3JQ2dg