Merge bk://kernel.bkbits.net/davem/net-2.6 into ppc970.osdl.org:/home/torvalds/v2.6/linux
2004/08/28 10:04:12-07:00 torvalds
Remove QIC-02 tape from Kconfig.

Let's leave the code for a while yet, and see if somebody screams.
But both Red Hat and SuSE claim it is unused, so it probably is. The
driver hasn't gotten updated in years, and google doesn't seem to
show any usage either.

BKrev: 4130bbedCFRI8CQBavpn4mwYbZ518A