[PATCH] ide: fix bad disk geometry hang
From: Alan Cox <alan@redhat.com>

If you get invalid geometry (eg a failed drive) the box hangs on boot because
we try and unregister the registered driver and end up in a locking mess.

We change the behaviour to simply leave the device unattached which avoids
the mess.

Signed-off-by: Bartlomiej Zolnierkiewicz <bzolnier@elka.pw.edu.pl>
Signed-off-by: Linus Torvalds <torvalds@osdl.org>

BKrev: 4130ccdfPbPJnM9-1PubdtWFozGglw