[PATCH] ppc32: fix the 'checkbin' target
The checkbin target on PPC32 isn't quite right.

First, one of the tests (to ensure that some instructions are known to
gas) is never actually invoked because 'checkbin' doesn't know about
stuff set in .config, so we always have the 'else' case run.  This
changes to always running the test and telling the user to upgrade to at
least binutils 2.12.1.

The next problem is that we were doing $(AS) -o /dev/null ...  in both
that test, as well as another.  The problem here is that the checkbin
target is run on the install targets, meaning that /dev/null will get
unlinked when the test passes.  To get around this we use .tmp_gas_check
as the output file instead.

Acked by Sam.

Signed-off-by: Tom Rini <trini@kernel.crashing.org>
Signed-off-by: Linus Torvalds <torvalds@osdl.org>

BKrev: 4136224cylZa383yRgLgBmzxgfqf4A