Merge davem@nuts.davemloft.net:/disk1/BK/net-2.6 into kernel.bkbits.net:/home/davem/net-2.6
2004/09/02 14:50:38-07:00 herbert
[IPV4/IPV6]: Use csum_sub() instead of csum_block_sub() w/zero 3rd arg.

Signed-off-by: Herbert Xu <herbert@gondor.apana.org.au>
Signed-off-by: David S. Miller <davem@davemloft.net>

2004/09/02 14:26:24-07:00 kaber
[IPV4/IPV6]: Fix suboptimal fragment sizing for last fragment

Yoshifuji's recent fragment patch prevents unnecessary fragmentation
when the data can be kept in a single packet, but only for the first
packet. When fragmenting, all fragments are still truncated to
multiples of 8 and we might end up creating an unnecessary fragment.

This dump shows the problem (MTU 1499):

172.16.1.123.32771 > 172.16.195.3.4135: udp 2937 (frag 7066:1472@0+)
172.16.1.123 > 172.16.195.3: udp (frag 7066:1472@1472+)
172.16.1.123 > 172.16.195.3: udp (frag 7066:1@2944)

This patch always builds mtu sized fragments and truncates the previous
fragment to a multiple of 8 bytes when allocating a new one. With the
patch the dump looks like this:


172.16.1.123.32772 > 172.16.195.3.4135: udp 2937 (frag 49641:1472@0+)
172.16.1.123 > 172.16.195.3: udp (frag 49641:1473@1472)

Signed-off-by: Patrick McHardy <kaber@trash.net>
Signed-off-by: David S. Miller <davem@davemloft.net>

2004/09/02 13:13:01-07:00 eric.lemoine
[NET]: Allow get/set of dev->weight via netlink.

Signed-off-by: David S. Miller <davem@davemloft.net>

2004/09/02 13:05:19-07:00 laforge
[IPV4]: Use nf_reset() in parp_redo().

Signed-off-by: Harald Welte <laforge@netfilter.org>
Signed-off-by: David S. Miller <davem@davemloft.net>

BKrev: 41380a58NY_9zBeBzOnY7zwGmXlxRg