[PATCH] urandom initialisation fix
The urandom driver returns the same thing after every boot.

The urandom pool is only cleared, and not initialized, and because there is
nothing in the primary pool, the reseeding is not successful.  The solution
is also the same, initialize not just the primary and secondary, but also
the urandom pool.

Signed-off-by: Andrew Morton <akpm@osdl.org>
Signed-off-by: Linus Torvalds <torvalds@osdl.org>

BKrev: 4138a944GgfkVOTE1TcV-UW-JIARpg