[PATCH] Root reservations for strict overcommit
This was on my TODO list for a while and it turns out someone already fixed
the armwaving overcommit mode for the same problem.  It is easy to get into
a situation where you have no overcommit and nothing can be done because
there is no memory to clean up the stable but non-useful state of the
machine.

The fix is trivial and duplicated from the armwaving overcommit code path.
The last 3% of the memory can be claimed by root processes only.  It isn't
a cure but it does seem to solve the real world problems - at least
providing you have enough memory for 3% to be useful 8).

Signed-off-by: Andrew Morton <akpm@osdl.org>
Signed-off-by: Linus Torvalds <torvalds@osdl.org>

BKrev: 4138a9bcNkehcxqL92KyW5l24yKGeg