Merge bk://linux-dj.bkbits.net/cpufreq into ppc970.osdl.org:/home/torvalds/v2.6/linux
2004/08/31 23:38:10+01:00 davej
[CPUFREQ] Fix cosmetic issue in powernow-k8 error handling

I tried to load the new SMP capable powernow k8 driver on a machine
with an too old BIOS. I got a lot of cryptic "init not cpu 0"
messages until the "MP systems not supported by PSB ..." message.

This patch switches the order of the checks around to print the
other message first. It will be still printed multiple times
because the cpufreq setup will probe all CPUs, but that's probably
ok.

From: Andi Kleen <ak@muc.de>
Signed-off-by: Dave Jones <davej@redhat.com>

2004/08/31 23:37:03+01:00 davej
[CPUFREQ] Remove duplicate information brought in with the new governor.

The current CPU frequency is already exported one directory upwards, so there
is no need to duplicate it in the ondemand governor.

From: Dominik Brodowski <linux@dominikbrodowski.de>
Signed-off-by: Dave Jones <davej@redhat.com>

2004/08/31 23:23:28+01:00 davej
[CPUFREQ] Fix up ARM drivers 'out of sync' problem.
The SA1100 and SA1110 platforms can handle situations well where the
CPU frequency is different to the value the cpufreq (and timing) code
thinks it is, e.g. when resuming from sleep. So, remove the flags
noting the opposite.

Signed-off-by: Dominik Brodowski <linux@dominikbrodowski.de>
Signed-off-by: Dave Jones <davej@redhat.com>

BKrev: 4138add75AtupUYxzJJDnHwdUbYuvg